requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
DateTime==4.3
dgl-cu102==0.7.1
idna==3.3
joblib==1.1.0
networkx==2.6.3
numpy==1.21.3
pandas==1.3.4
Pillow==8.4.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
scikit-learn==1.0
scipy==1.7.1
six==1.16.0
threadpoolctl==3.0.0
torch==1.7.1
torchaudio==0.7.2
torchvision==0.8.2
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.7
zope.interface==5.4.0
